wrapper: report progress for all disks
Report progress for all disks instead of reporting single 'progress'
attribute in state, that is reset to 0 for each copied disk. The
progress is reported in 'disks' attribute which is array of objects
containing 'path' and 'progress' keys.

It is possible (and expected, but not mandatory) to specify list of disk
paths in 'source_disks' atribute on input. The list is used to
initialize the state. Otherwise, the list of disks is build incrementally
during conversion.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>